# openapi.yaml
openapi: 3.0.3
info:
  title: AuthForge API
  description: JWT authentication & RBAC microservice with TOTP-based 2FA.
  version: 1.0.0
servers:
  - url: http://localhost:5000
paths:
...
  /api/admin/users:
    get:
      summary: List all users (admin only)
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: List of users }
        '403': { description: Insufficient role }
  /api/admin/users/{id}/role:
    patch:
      summary: Change a user's role (admin only)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role: { type: string, enum: [guest, user, admin] }
      responses:
        '200': { description: Role updated }
        '403': { description: Insufficient role }
        '404': { description: User not found }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT